ダウンロード不具合修正（int⇒long）

+++ dTVTerminalApp/app/src/main/cpp/src/DlnaDownload.cpp
@@ -102,7 +102,7 @@ bool DlnaDownload::start(DMP *dmp, JavaVM *vm, jobject object, jmethodID mid) {
     return d->start(vm, object);
 }
 
-void DlnaDownload::startDownload(std::string fileNameToSave, std::string dtcp1host_, int dtcp1port_, std::string url_, int cleartextSize, const char *xml) {
+void DlnaDownload::startDownload(std::string fileNameToSave, std::string dtcp1host_, int dtcp1port_, std::string url_, jlong cleartextSize, const char *xml) {
     if(fileNameToSave.empty() || dtcp1host_.empty() || url_.empty() || xml == nullptr){
         downloaderStatusHandler(DOWNLOADER_STATUS_ERROR_OCCURED, 0, this);
         return;
